+++ starsky/starskycore/Services/ReadMeta.cs
@@ -37,14 +37,17 @@ namespace starskycore.Services
 				return _readGpx.ReadGpxFromFileReturnAfterFirstField(_iStorage.ReadStream(fileIndexItemWithPath.FilePath));
 	        }
 
+	        // Read first the sidecar file
 	        var fileIndexItem = _readXmp.XmpGetSidecarFile(fileIndexItemWithPath);
 
 	        if ( fileIndexItem.IsoSpeed == 0 
 	             || string.IsNullOrEmpty(fileIndexItem.Make) 
 	             || fileIndexItem.DateTime.Year == 0)
 	        {
-		        var databaseItemFile = _readExif.ReadExifFromFile(fileIndexItemWithPath.FilePath);
-		        FileIndexCompareHelper.Compare(fileIndexItem, databaseItemFile);
+		        // so the sidecar file is not used
+		        var fileExifItemFile = _readExif.ReadExifFromFile(fileIndexItemWithPath.FilePath,fileIndexItemWithPath);
+		        // overwrite content with incomplete sidecar file (this file can contain tags)
+		        fileIndexItem = _readXmp.XmpGetSidecarFile(fileExifItemFile);
 	        }
 	        
             return fileIndexItem;