+++ history.md
@@ -17,6 +17,7 @@
 - [x]   Performance upgrade xmp/tiff files (does not check filehash again)
 - refactor xmp/exif module to support iStorage
 - bugfix to searching filehashes with null content
+- todo: write test with incomplete xmp file
 
 # version 0.1.5.1 - 2019-03-17
 - Breaking Change: added `LastEdited` field